requirement.txt
streamlit>=1.31.0

langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.7

faiss-cpu>=1.7.4

sentence-transformers>=2.2.2
huggingface-hub>=0.20.0

python-dotenv>=1.0.1

tiktoken>=0.6.0
